PR feedback changes implemented
